refactor(server): extract code delivery
