+++ crates/tidebreak-server/Cargo.toml
@@ -23,6 +23,7 @@ postgres = ["tidebreak-core/postgres", "sea-orm/sqlx-postgres", "tidebreak-core/
 [dependencies]
 tidebreak-shell-policy = { path = "../tidebreak-shell-policy" }
 tidebreak-code-execution = { path = "../tidebreak-code-execution" }
+tidebreak-code-delivery = { path = "../tidebreak-code-delivery" }
 tidebreak-code-remote = { path = "../tidebreak-code-remote" }
 tidebreak-harness = { path = "../tidebreak-harness" }
 tidebreak-managed-node = { path = "../tidebreak-managed-node" }
