+++ crates/tidebreak-server/src/code/delivery/api.rs
@@ -2,69 +2,106 @@
 
 use super::*;
 
-#[derive(Debug, Clone)]
-pub(super) enum DeliveryReader {
-    Gh(GhObservation),
-    Forge,
+#[derive(Clone)]
+pub(super) enum ServerDeliveryReader {
+    Gh {
+        observation: GhObservation,
+        runtime: Arc<CodeRuntime>,
+    },
+    Forge {
+        runtime: Arc<CodeRuntime>,
+        owner: OwnerId,
+    },
 }
 
-impl DeliveryReader {
-    pub(super) fn cache_scope(&self) -> &'static str {
+#[async_trait::async_trait]
+impl tidebreak_code_delivery::DeliveryReader for ServerDeliveryReader {
+    fn cache_scope(&self) -> &'static str {
         match self {
-            Self::Gh(_) => "gh",
-            Self::Forge => "forge-rest",
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(super) struct DeliveryAccess {
-    pub(super) capability: CodeGitHubCapability,
-    pub(super) reader: Option<DeliveryReader>,
-    pub(super) unavailable_kind: &'static str,
-}
-
-impl DeliveryAccess {
-    pub(super) fn source_error(&self) -> CodeDeliverySourceError {
-        CodeDeliverySourceError {
-            repository: None,
-            kind: self.unavailable_kind.into(),
-            message: self.capability.remediation.clone(),
-            retry_at: None,
+            Self::Gh { .. } => "gh",
+            Self::Forge { .. } => "forge-rest",
         }
     }
 
-    /// The reader, or the same refusal the list reads surface as a source
-    /// error — `gh_absent`/`gh_signed_out`/`gh_unavailable` on a local
-    /// machine, `git_forge_not_offered` with the connect path on a hosted
-    /// one.
-    pub(super) fn require_reader(&self) -> Result<DeliveryReader, ServerError> {
-        self.reader.clone().ok_or_else(|| {
-            ServerError::conflict_kind(self.unavailable_kind, self.capability.remediation.clone())
+    fn validate_pull_request_action(
+        &self,
+        action: &CodeDeliveryPullRequestAction,
+    ) -> Result<(), DeliveryError> {
+        if !matches!(self, Self::Forge { .. }) {
+            return Ok(());
+        }
+        match action {
+            CodeDeliveryPullRequestAction::MarkReady => Err(DeliveryError::conflict_kind(
+                "git_forge_mark_ready_unsupported",
+                "This hosted machine cannot mark a draft pull request ready because GitHub's pinned REST API does not expose that transition. Open the pull request on GitHub to mark it ready.",
+            )),
+            CodeDeliveryPullRequestAction::Merge { admin: true, .. } => {
+                Err(DeliveryError::conflict_kind(
+                    "git_forge_admin_merge_unsupported",
+                    "This hosted machine cannot request an admin branch-protection bypass through GitHub's stable REST API. Open the pull request on GitHub to merge with admin privileges.",
+                ))
+            }
+            _ => Ok(()),
+        }
+    }
+
+    async fn api(
+        &self,
+        target: &CodeGitHubRepositoryTarget,
+    ) -> Result<tidebreak_code_delivery::DeliveryApiHandle, String> {
+        let api = match self {
+            Self::Gh {
+                observation,
+                runtime,
+            } => ServerDeliveryApi::Gh {
+                observation: observation.clone(),
+                host: target.host.clone(),
+                search_path: runtime.gh_search_path_owned(),
+                runtime: Arc::clone(runtime),
+            },
+            Self::Forge { runtime, owner } => ServerDeliveryApi::Rest {
+                api_base: runtime.forge_api_base_for(&target.host),
+                credential: borrow_delivery_credential(runtime, owner, target).await?,
+                runtime: Arc::clone(runtime),
+            },
+        };
+        Ok(Arc::new(api))
+    }
+
+    async fn action_api(
+        &self,
+        target: &CodeGitHubRepositoryTarget,
+    ) -> Result<tidebreak_code_delivery::DeliveryApiHandle, DeliveryError> {
+        self.api(target).await.map_err(|message| match self {
+            Self::Forge { .. } => DeliveryError::conflict_kind("git_forge_not_offered", message),
+            Self::Gh { .. } => DeliveryError::bad_request_kind("github", message),
         })
     }
 }
 
 /// One authenticated Delivery transport. Reads and user actions select the
 /// same local `gh` or hosted forge REST path for each repository operation.
-pub(super) enum DeliveryApi {
+pub(super) enum ServerDeliveryApi {
     Gh {
         observation: GhObservation,
         host: String,
         search_path: Option<String>,
+        runtime: Arc<CodeRuntime>,
     },
     Rest {
         api_base: String,
         credential: GitCredential,
+        runtime: Arc<CodeRuntime>,
     },
 }
 
-impl DeliveryApi {
-    pub(super) fn can_mark_pull_request_ready(&self) -> bool {
+#[async_trait::async_trait]
+impl tidebreak_code_delivery::DeliveryApi for ServerDeliveryApi {
+    fn can_mark_pull_request_ready(&self) -> bool {
         matches!(self, Self::Gh { .. })
     }
 
-    pub(super) async fn get(&self, endpoint: &str) -> Result<Value, String> {
+    async fn get(&self, endpoint: &str) -> Result<Value, String> {
         match self {
             Self::Gh {
                 observation, host, ..
@@ -82,11 +119,186 @@ impl DeliveryApi {
             Self::Rest {
                 api_base,
                 credential,
+                ..
             } => crate::code::forge_rest::api_get(api_base, credential, endpoint).await,
         }
     }
 
-    pub(super) async fn merge_queue_membership(
+    async fn repository(&self, target: &CodeGitHubRepositoryTarget) -> Result<Value, String> {
+        match self {
+            Self::Gh {
+                observation, host, ..
+            } => {
+                let endpoint = format!("repos/{}/{}", target.owner, target.name);
+                run_api_json(
+                    observation
+                        .binary
+                        .as_deref()
+                        .expect("authenticated gh has a binary"),
+                    host,
+                    &endpoint,
+                )
+                .await
+            }
+            Self::Rest {
+                api_base,
+                credential,
+                ..
+            } => crate::code::forge_rest::repository(api_base, target, credential).await,
+        }
+    }
+
+    async fn pull_requests(
+        &self,
+        target: &CodeGitHubRepositoryTarget,
+        state: &str,
+        fields: &str,
+        checks_loaded: bool,
+        author: Option<&str>,
+    ) -> Result<Vec<Value>, String> {
+        match self {
+            Self::Gh { observation, .. } => {
+                let binary = observation
+                    .binary
+                    .as_deref()
+                    .expect("authenticated gh has a binary");
+                let repository = gh::cli_repository(&target.host, &target.owner, &target.name);
+                let limit = MAX_REMOTE_ITEMS_PER_REPO.to_string();
+                let mut args = vec![
+                    "pr",
+                    "list",
+                    "--repo",
+                    repository.as_str(),
+                    "--state",
+                    state,
+                    "--limit",
+                    limit.as_str(),
+                    "--json",
+                    fields,
+                ];
+                if let Some(author) = author {
+                    args.extend(["--author", author]);
+                }
+                let raw = gh::run_gh(Path::new("."), binary, &args, GH_READ_TIMEOUT).await?;
+                let value: Value = serde_json::from_str(&raw)
+                    .map_err(|error| format!("could not parse pull requests: {error}"))?;
+                Ok(value.as_array().cloned().unwrap_or_default())
+            }
+            Self::Rest {
+                api_base,
+                credential,
+                ..
+            } => {
+                crate::code::forge_rest::delivery_pull_requests(
+                    api_base,
+                    target,
+                    credential,
+                    state,
+                    checks_loaded,
+                )
+                .await
+            }
+        }
+    }
+
+    async fn deployments(&self, target: &CodeGitHubRepositoryTarget) -> Result<Value, String> {
+        match self {
+            Self::Gh {
+                observation, host, ..
+            } => {
+                let endpoint = format!(
+                    "repos/{}/{}/deployments?per_page=100",
+                    target.owner, target.name
+                );
+                run_api_json(
+                    observation
+                        .binary
+                        .as_deref()
+                        .expect("authenticated gh has a binary"),
+                    host,
+                    &endpoint,
+                )
+                .await
+            }
+            Self::Rest {
+                api_base,
+                credential,
+                ..
+            } => crate::code::forge_rest::deployments(api_base, target, credential).await,
+        }
+    }
+
+    async fn workflow_runs(
+        &self,
+        target: &CodeGitHubRepositoryTarget,
+        etag: Option<&str>,
+    ) -> Result<
+        tidebreak_code_delivery::EndpointRead<Vec<Value>>,
+        tidebreak_code_delivery::HostReadError,
+    > {
+        let read = match self {
+            Self::Gh {
+                observation,
+                runtime,
+                ..
+            } => {
+                let transport = crate::code::pr_fetch::FetchTransport::Gh {
+                    cwd: Path::new("."),
+                    binary: observation
+                        .binary
+                        .as_deref()
+                        .expect("authenticated gh has a binary"),
+                };
+                crate::code::pr_fetch::read_workflow_runs(
+                    &runtime.host_gate,
+                    transport,
+                    &target.host,
+                    &target.owner,
+                    &target.name,
+                    etag,
+                )
+                .await
+            }
+            Self::Rest {
+                api_base,
+                credential,
+                runtime,
+            } => {
+                let transport = crate::code::pr_fetch::FetchTransport::Rest {
+                    api_base,
+                    credential,
+                };
+                crate::code::pr_fetch::read_workflow_runs(
+                    &runtime.host_gate,
+                    transport,
+                    &target.host,
+                    &target.owner,
+                    &target.name,
+                    etag,
+                )
+                .await
+            }
+        };
+        match read {
+            Ok(crate::code::pr_fetch::EndpointRead::Fresh { value, etag }) => {
+                Ok(tidebreak_code_delivery::EndpointRead::Fresh { value, etag })
+            }
+            Ok(crate::code::pr_fetch::EndpointRead::NotModified) => {
+                Ok(tidebreak_code_delivery::EndpointRead::NotModified)
+            }
+            Ok(crate::code::pr_fetch::EndpointRead::Missing) => {
+                Ok(tidebreak_code_delivery::EndpointRead::Missing)
+            }
+            Err(crate::code::pr_fetch::FetchFailure::Parked(duration)) => {
+                Err(tidebreak_code_delivery::HostReadError::Parked(duration))
+            }
+            Err(error) => Err(tidebreak_code_delivery::HostReadError::Failed(
+                error.to_string(),
+            )),
+        }
+    }
+
+    async fn merge_queue_membership(
         &self,
         target: &CodeGitHubRepositoryTarget,
         number: u64,
@@ -119,6 +331,7 @@ impl DeliveryApi {
             Self::Rest {
                 api_base,
                 credential,
+                ..
             } => {
                 crate::code::forge_rest::merge_queue_state(api_base, target, credential, number)
                     .await
@@ -126,7 +339,7 @@ impl DeliveryApi {
         }
     }
 
-    pub(super) async fn pull_request(
+    async fn pull_request(
         &self,
         target: &CodeGitHubRepositoryTarget,
         repository: &CodeGitHubRepositoryRef,
@@ -162,6 +375,7 @@ impl DeliveryApi {
             Self::Rest {
                 api_base,
                 credential,
+                ..
             } => {
                 crate::code::forge_rest::delivery_pull_request(api_base, target, credential, number)
                     .await
@@ -169,11 +383,11 @@ impl DeliveryApi {
         }
     }
 
-    pub(super) async fn mark_pull_request_ready(
+    async fn mark_pull_request_ready(
         &self,
         target: &CodeGitHubRepositoryTarget,
         number: u64,
-    ) -> Result<(), ServerError> {
+    ) -> Result<(), DeliveryError> {
         match self {
             Self::Gh { search_path, .. } => gh::mark_pull_request_ready(
                 &target.host,
@@ -184,14 +398,14 @@ impl DeliveryApi {
             )
             .await
             .map_err(map_gh_error),
-            Self::Rest { .. } => Err(ServerError::conflict_kind(
+            Self::Rest { .. } => Err(DeliveryError::conflict_kind(
                 "git_forge_mark_ready_unsupported",
                 "This hosted machine cannot mark a draft pull request ready because GitHub's pinned REST API does not expose that transition. Open the pull request on GitHub to mark it ready.",
             )),
         }
     }
 
-    pub(super) async fn merge_pull_request(
+    async fn merge_pull_request(
         &self,
         target: &CodeGitHubRepositoryTarget,
         number: u64,
@@ -199,7 +413,7 @@ impl DeliveryApi {
         auto: bool,
         admin: bool,
         expected_head_sha: &str,
-    ) -> Result<(), ServerError> {
+    ) -> Result<(), DeliveryError> {
         match self {
             Self::Gh { search_path, .. } => gh::merge_pull_request_target(
                 &target.host,
@@ -217,6 +431,7 @@ impl DeliveryApi {
             Self::Rest {
                 api_base,
                 credential,
+                ..
             } => {
                 if auto {
                     return crate::code::forge_rest::enable_pull_request_auto_merge(
@@ -231,7 +446,7 @@ impl DeliveryApi {
                     .map_err(map_forge_action_error);
                 }
                 if admin {
-                    return Err(ServerError::conflict_kind(
+                    return Err(DeliveryError::conflict_kind(
                         "git_forge_admin_merge_unsupported",
                         "This hosted machine cannot request an admin branch-protection bypass through GitHub's stable REST API. Open the pull request on GitHub to merge with admin privileges.",
                     ));
@@ -250,11 +465,11 @@ impl DeliveryApi {
         }
     }
 
-    pub(super) async fn create_stack(
+    async fn create_stack(
         &self,
         target: &CodeGitHubRepositoryTarget,
         numbers: &[u64],
-    ) -> Result<(), ServerError> {
+    ) -> Result<(), DeliveryError> {
         match self {
             Self::Gh {
                 host, search_path, ..
@@ -270,18 +485,19 @@ impl DeliveryApi {
             Self::Rest {
                 api_base,
                 credential,
+                ..
             } => crate::code::forge_rest::create_stack(api_base, target, credential, numbers)
                 .await
                 .map_err(map_forge_action_error),
         }
     }
 
-    pub(super) async fn update_pull_request_state(
+    async fn update_pull_request_state(
         &self,
         target: &CodeGitHubRepositoryTarget,
         number: u64,
         state: &str,
-    ) -> Result<(), ServerError> {
+    ) -> Result<(), DeliveryError> {
         match (self, state) {
             (Self::Gh { search_path, .. }, "closed") => gh::close_pull_request_target(
                 &target.host,
@@ -301,13 +517,14 @@ impl DeliveryApi {
             )
             .await
             .map_err(map_gh_error),
-            (Self::Gh { .. }, _) => Err(ServerError::internal(
+            (Self::Gh { .. }, _) => Err(DeliveryError::internal(
                 "Delivery requested an unsupported pull request state",
             )),
             (
                 Self::Rest {
                     api_base,
                     credential,
+                    ..
                 },
                 state,
             ) => crate::code::forge_rest::update_pull_request_state(
@@ -318,12 +535,12 @@ impl DeliveryApi {
         }
     }
 
-    pub(super) async fn comment_on_pull_request(
+    async fn comment_on_pull_request(
         &self,
         target: &CodeGitHubRepositoryTarget,
         number: u64,
         body: &str,
-    ) -> Result<(), ServerError> {
+    ) -> Result<(), DeliveryError> {
         match self {
             Self::Gh { search_path, .. } => gh::comment_on_pull_request_target(
                 &target.host,
@@ -338,6 +555,7 @@ impl DeliveryApi {
             Self::Rest {
                 api_base,
                 credential,
+                ..
             } => crate::code::forge_rest::comment_on_pull_request(
                 api_base, target, credential, number, body,
             )
@@ -346,11 +564,11 @@ impl DeliveryApi {
         }
     }
 
-    pub(super) async fn rerun_failed_jobs(
+    async fn rerun_failed_jobs(
         &self,
         target: &CodeGitHubRepositoryTarget,
         run_id: u64,
-    ) -> Result<(), ServerError> {
+    ) -> Result<(), DeliveryError> {
         match self {
             Self::Gh { observation, .. } => gh::rerun_failed_jobs_with_observation(
                 observation,
@@ -364,17 +582,18 @@ impl DeliveryApi {
             Self::Rest {
                 api_base,
                 credential,
+                ..
             } => crate::code::forge_rest::rerun_failed_jobs(api_base, target, credential, run_id)
                 .await
                 .map_err(map_forge_action_error),
         }
     }
 
-    pub(super) async fn rerun_workflow(
+    async fn rerun_workflow(
         &self,
         target: &CodeGitHubRepositoryTarget,
         run_id: u64,
-    ) -> Result<(), ServerError> {
+    ) -> Result<(), DeliveryError> {
         match self {
             Self::Gh { observation, .. } => gh::rerun_workflow_with_observation(
                 observation,
@@ -388,88 +607,10 @@ impl DeliveryApi {
             Self::Rest {
                 api_base,
                 credential,
+                ..
             } => crate::code::forge_rest::rerun_workflow(api_base, target, credential, run_id)
                 .await
                 .map_err(map_forge_action_error),
-        }
-    }
-}
-
-pub(super) async fn delivery_api(
-    runtime: &CodeRuntime,
-    owner: &OwnerId,
-    reader: &DeliveryReader,
-    target: &CodeGitHubRepositoryTarget,
-) -> Result<DeliveryApi, String> {
-    match reader {
-        DeliveryReader::Gh(observation) => Ok(DeliveryApi::Gh {
-            observation: observation.clone(),
-            host: target.host.clone(),
-            search_path: runtime.gh_search_path_owned(),
-        }),
-        DeliveryReader::Forge => {
-            let credential = borrow_delivery_credential(runtime, owner, target).await?;
-            Ok(DeliveryApi::Rest {
-                api_base: runtime.forge_api_base_for(&target.host),
-                credential,
-            })
-        }
-    }
-}
-
-/// Build the same transport as reads, while preserving the forge-specific
-/// refusal kind if a credential mint fails after the availability probe.
-///
-/// The probe and mint are separate gateway calls. A disconnect or expired
-/// caller session between them must remain a hosted-forge refusal, not turn
-/// into a generic GitHub request error.
-pub(super) async fn delivery_action_api(
-    runtime: &CodeRuntime,
-    owner: &OwnerId,
-    reader: &DeliveryReader,
-    target: &CodeGitHubRepositoryTarget,
-) -> Result<DeliveryApi, ServerError> {
-    delivery_api(runtime, owner, reader, target)
-        .await
-        .map_err(|message| {
-            if matches!(reader, DeliveryReader::Forge) {
-                ServerError::conflict_kind("git_forge_not_offered", message)
-            } else {
-                ServerError::bad_request_kind("github", message)
-            }
-        })
-}
-
-pub(super) async fn resolve_repository_for_api(
-    runtime: &CodeRuntime,
-    api: &DeliveryApi,
-    target: &CodeGitHubRepositoryTarget,
-    tidebreak_repo_id: Option<tidebreak_core::RepoId>,
-    force_refresh: bool,
-) -> Result<CodeGitHubRepositoryRef, String> {
-    match api {
-        DeliveryApi::Gh { observation, .. } => {
-            resolve_repository_cached(
-                runtime,
-                observation
-                    .binary
-                    .as_deref()
-                    .expect("authenticated gh has a binary"),
-                target,
-                tidebreak_repo_id,
-                force_refresh,
-            )
-            .await
-        }
-        DeliveryApi::Rest { credential, .. } => {
-            resolve_repository_rest_cached(
-                runtime,
-                target,
-                tidebreak_repo_id,
-                force_refresh,
-                credential,
-            )
-            .await
         }
     }
 }
@@ -489,10 +630,10 @@ pub(super) fn github_capability(observation: &GhObservation) -> CodeGitHubCapabi
 /// probe is the source of availability and identity. Every other machine
 /// keeps the existing GitHub CLI observation unchanged.
 pub(super) async fn delivery_access(
-    runtime: &CodeRuntime,
+    runtime: &Arc<CodeRuntime>,
     owner: &OwnerId,
     force_refresh: bool,
-) -> DeliveryAccess {
+) -> tidebreak_code_delivery::DeliveryAccess {
     if let Some(lender) = runtime.git_credentials() {
         return match lender.git_forge_identity(owner).await {
             Ok(identity) => {
@@ -500,18 +641,21 @@ pub(super) async fn delivery_access(
                     GitForgeAttribution::Person { login, .. } => Some(login),
                     GitForgeAttribution::Bot { bot_login } => bot_login,
                 };
-                DeliveryAccess {
+                tidebreak_code_delivery::DeliveryAccess {
                     capability: CodeGitHubCapability {
                         found: true,
                         authenticated: Some(true),
                         viewer_login,
                         remediation: String::new(),
                     },
-                    reader: Some(DeliveryReader::Forge),
+                    reader: Some(Arc::new(ServerDeliveryReader::Forge {
+                        runtime: Arc::clone(runtime),
+                        owner: owner.clone(),
+                    })),
                     unavailable_kind: "git_forge_not_offered",
                 }
             }
-            Err(refusal) => DeliveryAccess {
+            Err(refusal) => tidebreak_code_delivery::DeliveryAccess {
                 capability: CodeGitHubCapability {
                     found: !matches!(&refusal, crate::obo_gateway::GitForgeError::NoGitForge),
                     authenticated: Some(false),
@@ -531,9 +675,13 @@ pub(super) async fn delivery_access(
         gh::observe_gh(search_path.as_deref()).await
     };
     let unavailable_kind = observation_error_kind(&observation);
-    let reader =
-        (observation.authenticated == Some(true)).then(|| DeliveryReader::Gh(observation.clone()));
-    DeliveryAccess {
+    let reader = (observation.authenticated == Some(true)).then(|| {
+        Arc::new(ServerDeliveryReader::Gh {
+            observation: observation.clone(),
+            runtime: Arc::clone(runtime),
+        }) as tidebreak_code_delivery::DeliveryReaderHandle
+    });
+    tidebreak_code_delivery::DeliveryAccess {
         capability: github_capability(&observation),
         reader,
         unavailable_kind,
@@ -541,7 +689,7 @@ pub(super) async fn delivery_access(
 }
 
 /// Borrow one credential for one repository Delivery operation.
-pub(super) async fn borrow_delivery_credential(
+async fn borrow_delivery_credential(
     runtime: &CodeRuntime,
     owner: &OwnerId,
     target: &CodeGitHubRepositoryTarget,
