+++ crates/tidebreak-server/src/code/reconcile.rs
@@ -18,12 +18,15 @@ use tracing::debug;
 use tidebreak_core::db::code::{
     list_fact_repo_identities_all_owners, list_repos_all_owners, set_repo_origin,
 };
-use tidebreak_core::{CodePullRequestFact, CodePullRequestId, CodePullRequestState, OwnerId};
+use tidebreak_core::OwnerId;
 
 use super::delivery::{query_pull_requests, repository_target_from_local, MAX_REPOSITORIES};
 use super::runtime::CodeRuntime;
-use crate::routes::code::types::{
-    CodeDeliveryPullRequestQuery, CodeDeliveryPullRequestSummary, CodeGitHubRepositoryTarget,
+use crate::routes::code::types::{CodeDeliveryPullRequestQuery, CodeGitHubRepositoryTarget};
+#[cfg(test)]
+pub(crate) use tidebreak_code_delivery::{
+    StackParentCandidate, StackParentEdge, StackParentIndex, StackParentResolution,
+    StackParentUnresolvedReason, StackPullRequestIdentity, StackRepositoryIdentity,
 };
 
 /// Coprime with the 47s watch and 53s trigger sweeps, so three GitHub-reading
@@ -243,217 +246,6 @@ pub(crate) async fn sweep_reconcile(runtime: &Arc<CodeRuntime>) {
             debug!("reconcile read failed: {err:?}");
         }
         super::delivery::refresh_workflow_runs(runtime, &owner, &deduped).await;
-    }
-}
-
-/// Build a durable fact from one parsed delivery summary.
-///
-/// `None` when the summary's state token is not one the fact vocabulary
-/// stores — the delivery parser already lowercases and normalizes a merged
-/// close, so an unknown token here is a contract drift worth staying quiet
-/// about rather than guessing.
-pub(crate) fn fact_from_summary(
-    owner: &OwnerId,
-    summary: &CodeDeliveryPullRequestSummary,
-    now: chrono::DateTime<chrono::Utc>,
-) -> Option<CodePullRequestFact> {
-    let state = CodePullRequestState::from_str(&summary.state)?;
-    Some(CodePullRequestFact {
-        id: CodePullRequestId::new(),
-        owner: owner.clone(),
-        host: summary.repository.host.clone(),
-        repo_owner: summary.repository.owner.clone(),
-        repo_name: summary.repository.name.clone(),
-        number: summary.number,
-        url: summary.url.clone(),
-        title: summary.title.clone(),
-        state,
-        draft: summary.draft,
-        author: summary.author.clone(),
-        head_branch: summary.head_branch.clone(),
-        base_branch: summary.base_branch.clone(),
-        head_sha: summary.head_sha.clone(),
-        created_at: summary.created_at,
-        updated_at: summary.updated_at,
-        merged_at: summary.merged_at,
-        closed_at: summary.closed_at,
-        first_seen_at: now,
-        last_seen_at: now,
-        live: None,
-    })
-}
-
-/// Case-insensitive repository identity used while resolving stack edges.
-///
-/// GitHub repository owner and name tokens are case-insensitive. Branch names
-/// remain case-sensitive and live on [`StackParentEdge`].
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub(crate) struct StackRepositoryIdentity {
-    pub(crate) host: String,
-    pub(crate) owner: String,
-    pub(crate) name: String,
-}
-
-impl StackRepositoryIdentity {
-    pub(crate) fn new(host: &str, owner: &str, name: &str) -> Option<Self> {
-        let host = host.trim().trim_end_matches('.').to_ascii_lowercase();
-        let owner = owner.trim().to_ascii_lowercase();
-        let name = name.trim().to_ascii_lowercase();
-        let name = name.strip_suffix(".git").unwrap_or(&name).to_owned();
-        if host.is_empty() || owner.is_empty() || name.is_empty() {
-            return None;
-        }
-        Some(Self { host, owner, name })
-    }
-}
-
-/// Immutable pull-request identity after a stack edge resolves.
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub(crate) struct StackPullRequestIdentity {
-    pub(crate) base_repository: StackRepositoryIdentity,
-    pub(crate) number: u64,
-}
-
-/// The mutable branch edge that may resolve to an immutable pull request.
-///
-/// The base repository scopes the pull-request number. The head repository
-/// distinguishes same-named branches in forks of that base repository.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct StackParentEdge {
-    pub(crate) base_repository: StackRepositoryIdentity,
-    pub(crate) head_repository: Option<StackRepositoryIdentity>,
-    pub(crate) head_branch: String,
-}
-
-impl StackParentEdge {
-    pub(crate) fn new(
-        base_repository: StackRepositoryIdentity,
-        head_repository: Option<StackRepositoryIdentity>,
-        head_branch: &str,
-    ) -> Option<Self> {
-        if head_branch.is_empty() {
-            return None;
-        }
-        Some(Self {
-            base_repository,
-            head_repository,
-            head_branch: head_branch.to_owned(),
-        })
-    }
-}
-
-/// One possible open stack parent from a host observation or durable fact.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct StackParentCandidate {
-    pub(crate) pull_request: StackPullRequestIdentity,
-    pub(crate) open: bool,
-    pub(crate) head_repository: Option<StackRepositoryIdentity>,
-    pub(crate) head_branch: Option<String>,
-}
-
-/// Why a branch edge could not safely resolve to one pull request.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum StackParentUnresolvedReason {
-    MissingParent,
-    IncompleteHostIdentity,
-    Ambiguous {
-        candidates: Vec<StackPullRequestIdentity>,
-    },
-}
-
-/// A stack edge resolves to one immutable pull request or stays explicit.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum StackParentResolution {
-    Resolved(StackPullRequestIdentity),
-    Unresolved {
-        edge: StackParentEdge,
-        reason: StackParentUnresolvedReason,
-    },
-}
-
-/// Open pull-request heads indexed by full fork-qualified branch identity.
-#[derive(Debug, Default)]
-pub(crate) struct StackParentIndex {
-    exact: HashMap<StackParentEdge, Vec<StackPullRequestIdentity>>,
-    incomplete_branches: HashSet<(StackRepositoryIdentity, String)>,
-    incomplete_repositories: HashSet<StackRepositoryIdentity>,
-}
-
-impl StackParentIndex {
-    pub(crate) fn new(candidates: impl IntoIterator<Item = StackParentCandidate>) -> Self {
-        let mut index = Self::default();
-        for candidate in candidates {
-            if !candidate.open {
-                continue;
-            }
-            match (candidate.head_repository, candidate.head_branch) {
-                (Some(head_repository), Some(head_branch)) if !head_branch.is_empty() => {
-                    let edge = StackParentEdge {
-                        base_repository: candidate.pull_request.base_repository.clone(),
-                        head_repository: Some(head_repository),
-                        head_branch,
-                    };
-                    index
-                        .exact
-                        .entry(edge)
-                        .or_default()
-                        .push(candidate.pull_request);
-                }
-                (None, Some(head_branch)) if !head_branch.is_empty() => {
-                    index
-                        .incomplete_branches
-                        .insert((candidate.pull_request.base_repository, head_branch));
-                }
-                _ => {
-                    index
-                        .incomplete_repositories
-                        .insert(candidate.pull_request.base_repository);
-                }
-            }
-        }
-        for candidates in index.exact.values_mut() {
-            candidates.sort();
-            candidates.dedup();
-        }
-        index
-    }
-
-    pub(crate) fn resolve(
-        &self,
-        edge: &StackParentEdge,
-        child: Option<&StackPullRequestIdentity>,
-    ) -> StackParentResolution {
-        if edge.head_repository.is_none() {
-            return StackParentResolution::Unresolved {
-                edge: edge.clone(),
-                reason: StackParentUnresolvedReason::IncompleteHostIdentity,
-            };
-        }
-        let mut candidates = self.exact.get(edge).cloned().unwrap_or_default();
-        if let Some(child) = child {
-            candidates.retain(|candidate| candidate != child);
-        }
-        let incomplete = self.incomplete_repositories.contains(&edge.base_repository)
-            || self
-                .incomplete_branches
-                .contains(&(edge.base_repository.clone(), edge.head_branch.clone()));
-        let reason = if incomplete {
-            StackParentUnresolvedReason::IncompleteHostIdentity
-        } else if candidates.is_empty() {
-            StackParentUnresolvedReason::MissingParent
-        } else if candidates.len() == 1 {
-            return StackParentResolution::Resolved(
-                candidates
-                    .pop()
-                    .expect("a one-candidate stack edge has a parent"),
-            );
-        } else {
-            StackParentUnresolvedReason::Ambiguous { candidates }
-        };
-        StackParentResolution::Unresolved {
-            edge: edge.clone(),
-            reason,
-        }
     }
 }
