+++ crates/tidebreak-server/src/tests/code_git.rs
@@ -839,7 +839,7 @@ async fn a_hosted_delivery_page_reads_and_acts_over_forge_rest() {
     );
     assert_eq!(
         lender.minted(),
-        vec!["acme/demo".to_owned(); 15],
+        vec!["acme/demo".to_owned(); borrows_before_unsupported],
         "every read and action borrows only for the registered repository"
     );
 }
